+++ src/vmecpp/cpp/vmecpp/common/CMakeLists.txt
@@ -8,7 +8,6 @@ add_subdirectory(magnetic_configuration_lib)
 add_subdirectory(magnetic_field_provider)
 add_subdirectory(makegrid_lib)
 add_subdirectory(sizes)
-add_subdirectory(spectral_width)
 add_subdirectory(util)
 add_subdirectory(vmec_indata)
 set (vmecpp_sources "${vmecpp_sources}" PARENT_SCOPE)
