--- src/vmecpp/cpp/vmecpp/common/spectral_width/CMakeLists.txt
@@ -1,5 +1,0 @@
-list (APPEND vmecpp_sources
-  ${CMAKE_CURRENT_SOURCE_DIR}/spectral_width.cc
-  ${CMAKE_CURRENT_SOURCE_DIR}/spectral_width.h
-)
-set (vmecpp_sources "${vmecpp_sources}" PARENT_SCOPE)
